requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
mysql-connector==2.2.9
numpy==1.16.2
pandas==0.24.2
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
